requirements.txt
dspy-ai
dotenv